refactor: distribute code base to appropriate .py files and remove deprecated functions
