+++ patternProcessing.py
@@ -1,6 +1,5 @@
 import re
 
-
 def parse_pattern(s_t, pat):
     p = pat.copy()
     pattern_arr = []
@@ -39,3 +38,14 @@ def build_type(inp_string):
             inp_type += "L"
             inp_string = re.sub(r'^[a-z]', "", inp_string, 1)
     return inp_type
+
+
+def is_valid_seq(seq_mat):
+    seq_list = seq_mat.copy().flatten()
+    seq_type = ""
+    for pos in range(len(seq_list)):
+        if seq_type == "":
+            seq_type = build_type(seq_list[pos])
+        elif seq_type != build_type(seq_list[pos]):
+            return False
+    return seq_type
